requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
openai==1.3.9
python-dotenv==1.0.0

